# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FacilityTrack Aquatic Operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /test-results:
    get:
      operationId: listTestResults
      tags: [test-results]
      summary: List test results
      parameters:
        - name: poolId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: isCompliant
          in: query
          schema:
            type: ["boolean", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of test results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestResult"
    post:
      operationId: createTestResult
      tags: [test-results]
      summary: Record a pool water test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestResultBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
  /test-results/{id}:
    get:
      operationId: getTestResult
      tags: [test-results]
      summary: Get a test result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
        "404":
          description: Not found
    patch:
      operationId: updateTestResult
      tags: [test-results]
      summary: Update a test result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTestResultBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
    delete:
      operationId: deleteTestResult
      tags: [test-results]
      summary: Delete a test result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── WATER BALANCE ────────────────────────────────────────────────────────────
  /water-balance-tests:
    get:
      operationId: listWaterBalanceTests
      tags: [water-balance]
      summary: List water balance tests
      parameters:
        - name: poolId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of water balance tests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WaterBalanceTest"
    post:
      operationId: createWaterBalanceTest
      tags: [water-balance]
      summary: Record a water balance test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWaterBalanceTestBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaterBalanceTest"
  /water-balance-tests/{id}:
    get:
      operationId: getWaterBalanceTest
      tags: [water-balance]
      summary: Get a water balance test
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Water balance test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaterBalanceTest"
        "404":
          description: Not found
    patch:
      operationId: updateWaterBalanceTest
      tags: [water-balance]
      summary: Update a water balance test
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWaterBalanceTestBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaterBalanceTest"
    delete:
      operationId: deleteWaterBalanceTest
      tags: [water-balance]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── POOL CLOSURES ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TestResult:
      type: object
      properties:
        id:
          type: integer
        poolId:
          type: integer
        poolName:
          type: ["string", "null"]
        testedBy:
          type: ["string", "null"]
        testedAt:
          type: string
          format: date-time
        freeChlorine:
          type: ["number", "null"]
        totalAvailableChlorine:
          type: ["number", "null"]
        combinedChlorine:
          type: ["number", "null"]
        ph:
          type: ["number", "null"]
        temperature:
          type: ["number", "null"]
        turbidity:
          type: ["number", "null"]
        totalAlkalinity:
          type: ["number", "null"]
        isCompliant:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - poolId
        - isCompliant
        - testedAt
        - createdAt
    CreateTestResultBody:
      type: object
      properties:
        poolId:
          type: integer
        testedBy:
          type: ["string", "null"]
        testedAt:
          type: ["string", "null"]
        freeChlorine:
          type: ["number", "null"]
        totalAvailableChlorine:
          type: ["number", "null"]
        combinedChlorine:
          type: ["number", "null"]
        ph:
          type: ["number", "null"]
        temperature:
          type: ["number", "null"]
        turbidity:
          type: ["number", "null"]
        totalAlkalinity:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - poolId
    UpdateTestResultBody:
      type: object
      properties:
        freeChlorine:
          type: ["number", "null"]
        totalAvailableChlorine:
          type: ["number", "null"]
        combinedChlorine:
          type: ["number", "null"]
        ph:
          type: ["number", "null"]
        temperature:
          type: ["number", "null"]
        turbidity:
          type: ["number", "null"]
        totalAlkalinity:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]

    # ── WATER BALANCE TEST ────────────────────────────────────────────────────
    WaterBalanceTest:
      type: object
      properties:
        id:
          type: integer
        poolId:
          type: integer
        poolName:
          type: ["string", "null"]
        testedBy:
          type: ["string", "null"]
        testedAt:
          type: string
          format: date-time
        ph:
          type: ["number", "null"]
        totalAlkalinity:
          type: ["number", "null"]
        calciumHardness:
          type: ["number", "null"]
        cyanuricAcid:
          type: ["number", "null"]
        totalDissolvedSolids:
          type: ["number", "null"]
        langelier:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - poolId
        - testedAt
        - createdAt
    CreateWaterBalanceTestBody:
      type: object
      properties:
        poolId:
          type: integer
        testedBy:
          type: ["string", "null"]
        testedAt:
          type: ["string", "null"]
        ph:
          type: ["number", "null"]
        totalAlkalinity:
          type: ["number", "null"]
        calciumHardness:
          type: ["number", "null"]
        cyanuricAcid:
          type: ["number", "null"]
        totalDissolvedSolids:
          type: ["number", "null"]
        langelier:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - poolId
    UpdateWaterBalanceTestBody:
      type: object
      properties:
        ph:
          type: ["number", "null"]
        totalAlkalinity:
          type: ["number", "null"]
        calciumHardness:
          type: ["number", "null"]
        cyanuricAcid:
          type: ["number", "null"]
        totalDissolvedSolids:
          type: ["number", "null"]
        langelier:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]

    # ── POOL CLOSURE ──────────────────────────────────────────────────────────
...